Move create_glue_database function into glue_utils
