+++ containers/daap-python-base/src/var/task/glue_utils.py
@@ -5,6 +5,24 @@ from data_platform_logging import DataPlatformLogger
 glue_client = boto3.client("glue")
 
 
+def create_glue_database(glue_client, database_name, logger):
+    """If a glue database doesn't exist, create a glue database"""
+    try:
+        glue_client.get_database(Name=database_name)
+    except ClientError as e:
+        if e.response["Error"]["Code"] == "EntityNotFoundException":
+            db_meta = {
+                "DatabaseInput": {
+                    "Description": "database for {} products".format(database_name),
+                    "Name": database_name,
+                }
+            }
+            glue_client.create_database(**db_meta)
+        else:
+            logger.error("Unexpected error: %s" % e)
+            raise
+
+
 def delete_glue_table(
     data_product_name: str, table_name: str, logger: DataPlatformLogger
 ) -> str | None:
